Feat(item): 상품 수정

+++ backend/yam/src/main/java/com/ssafy/yam/domain/item/repository/ItemRepository.java
@@ -3,11 +3,13 @@ package com.ssafy.yam.domain.item.repository;
 import com.ssafy.yam.domain.item.entity.Item;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
 
-public interface ItemRepository extends JpaRepository<Item, Long> {
+public interface ItemRepository extends JpaRepository<Item, Integer> {
     Item findItemByItemId(int itemId);
     List<Item> findAllBy(Pageable pageable);
-//    List<Item> findAllBy();
+
 }
